fully protected authorization success

+++ backend/controllers/goalContoller.js
@@ -1,9 +1,10 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // Read request
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -14,7 +15,10 @@ const postGoal = asyncHandler(async (req, res) => {
     throw new Error("please add a text fiels");
   }
 
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id
+  });
   res.status(200).json(goal);
 });
 
@@ -25,6 +29,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+  if (!user) {
+    res.status(401)
+    throw new Error("User not found!")
+  }
+
+  //make sure the logged in user matches to the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error("User not authorized!")
   }
 
   const newGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -41,6 +59,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+  if (!user) {
+    res.status(401)
+    throw new Error("User not found!")
+  }
+
+  //make sure the logged in user matches to the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error("User not authorized!")
   }
 
   await Goal.findByIdAndDelete(req.params.id);
